Update gem dependencies
`bundle update`
